Start to separate
