+++ tsconfig.json
@@ -7,6 +7,7 @@
       "@layouts/*": ["src/layouts/*"],
       "@stores/*": ["src/stores/*"],
       "@utils/*": ["src/utils/*"],
+      "@lib/*": ["src/lib/*"],
     }
   }
 }
